feat: move app out of internal
