+++ cmd/root.go
@@ -5,8 +5,8 @@ import (
 	"log/slog"
 	"os"
 
+	"github.com/masamerc/sevp/app"
 	"github.com/masamerc/sevp/internal"
-	"github.com/masamerc/sevp/internal/app"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
